+++ src/LoomSlotter.cpp
@@ -2,7 +2,7 @@
 
 using namespace LoomNet;
 
-Slotter::Slotter(const uint8_t send_slot, const uint8_t total_slots, const uint8_t cycles_per_refresh, const uint8_t cycle_gap, const uint8_t batch_gap, const uint8_t send_count, const uint8_t recv_slot, const uint8_t recv_count)
+Slotter::Slotter(const uint8_t send_slot, const uint16_t total_slots, const uint8_t cycles_per_refresh, const uint8_t cycle_gap, const uint8_t batch_gap, const uint8_t send_count, const uint8_t recv_slot, const uint8_t recv_count)
 	: m_send_slot(send_slot)
 	, m_send_count(send_count)
 	, m_recv_slot(recv_slot)
